lang_chain_poc/requirements.txt
# === LangChain POC — Databricks pinned requirements ===
# Tip: after %pip install, run %restart_python

python-dotenv==1.2.1
# Core LangChain stack
langchain==0.3.27
langchain-core==0.3.79
langchain-community==0.3.31
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langchain-pinecone==0.2.12
langchain-tavily==0.2.12

# LLM + embeddings client
openai==2.6.1
tiktoken==0.12.0

# Vector DB
pinecone-client==6.0.0
pinecone==7.3.0

# Utilities
pydantic>=2.10.6,<3
beautifulsoup4==4.12.3
requests==2.32.5
tqdm==4.67.1
rich==13.9.4

# Search tools (choose what you use; both can be installed safely)
tavily-python==0.7.12
google-search-results==2.4.2

# Optional direct pins to match the resolved environment and speed up pip solver
SQLAlchemy==2.0.44
aiohttp==3.13.2
httpx==0.28.1
httpx-sse==0.4.3
pydantic-settings==2.11.0
langsmith==0.4.38
dataclasses-json==0.6.7

huggingface_hub==0.25.2
requests==2.32.5
pinecone==7.3.0
pinecone-client==6.0.0
langchain==0.3.27
langchain-core==0.3.79
langchain-text-splitters==0.3.11
langchain-pinecone==0.2.12
lxml==6.0.2

alpha_vantage
yfinance

